+++ src/layouts/AccessWalletLayout/components/HardwareModal/HardwareModal.vue
@@ -54,7 +54,8 @@ import trezorHov from '@/assets/images/icons/button-trezor-hover.png';
 import keepkey from '@/assets/images/icons/button-keepkey.png';
 import keepkeyHov from '@/assets/images/icons/button-keepkey-hover.png';
 import WalletOption from '../WalletOption';
-import { ErrorHandler, Misc } from '@/helpers';
+import { Toast, Misc } from '@/helpers';
+import { isSupported } from 'u2f-api';
 import {
   LedgerWallet,
   KeepkeyWallet,
@@ -81,40 +82,36 @@ export default {
     return {
       selected: '',
       mayNotBeAttached: false,
+      isU2FSupported: false,
       items: [
         {
           name: 'ledger',
           imgPath: ledger,
           imgHoverPath: ledgerHov,
           text: 'Ledger',
-          disabled:
-            window.location.protocol === 'https:' &&
-            Misc.browserName() !== 'chrome'
+          disabled: false
         },
         {
           name: 'bitbox',
           imgPath: bitbox,
           imgHoverPath: bitboxHov,
           text: 'Digital Bitbox',
-          disabled:
-            window.location.protocol === 'https:' &&
-            Misc.browserName() !== 'chrome'
+          disabled: false
         },
         {
           name: 'secalot',
           imgPath: secalot,
           imgHoverPath: secalotHov,
           text: 'Secalot',
-          disabled:
-            window.location.protocol === 'https:' &&
-            Misc.browserName() !== 'chrome'
+          disabled: false
         },
         {
           name: 'trezor',
           imgPath: trezor,
           imgHoverPath: trezorHov,
           text: 'Trezor',
-          disabled: false
+          disabled:
+            Misc.browserName() !== 'chrome' && Misc.browserName() !== 'firefox'
         },
         {
           name: 'keepkey',
@@ -129,11 +126,31 @@ export default {
     };
   },
   mounted() {
+    isSupported().then(res => {
+      this.items.forEach(item => {
+        const u2fhw = ['secalot', 'ledger', 'bitbox'];
+        const inMobile = ['secalot', 'keepkey'];
+
+        if (u2fhw.includes(item.name))
+          item.disabled = !(
+            (Misc.browserName() === 'chrome' ||
+              Misc.browserName() === 'opera') &&
+            res
+          );
+        if (this.isMobile()) item.disabled = !inMobile.includes(item.name);
+      });
+    });
     this.$refs.hardware.$on('hidden', () => {
       this.selected = '';
     });
   },
   methods: {
+    isMobile() {
+      return (
+        typeof window.orientation !== 'undefined' ||
+        navigator.userAgent.indexOf('IEMobile') !== -1
+      );
+    },
     continueAccess() {
       const showPluggedInReminder = setTimeout(() => {
         this.mayNotBeAttached = true;
@@ -175,7 +192,10 @@ export default {
             .catch(KeepkeyWallet.errorHandler);
           break;
         default:
-          ErrorHandler(new Error('No switch address for given account.'), true);
+          Toast.responseHandler(
+            new Error('No switch address for given account.'),
+            Toast.ERROR
+          );
           break;
       }
     },