+++ src/layouts/InterfaceLayout/InterfaceLayout.vue
@@ -27,7 +27,7 @@
     />
     <print-modal
       ref="printModal"
-      :priv-key="wallet.privateKey"
+      :priv-key="!wallet.isPubOnly"
       :address="account.address"
     />
     <address-qrcode-modal ref="addressQrcodeModal" :address="account.address" />
@@ -37,7 +37,6 @@
     <div class="mobile-interface-address-block">
       <mobile-interface-address
         :address="address"
-        :trigger-alert="triggerAlert"
         :print="print"
         :switch-addr="switchAddress"
       />
@@ -55,18 +54,10 @@
         </div>
       </div>
       <div class="contents">
-        <b-alert
-          :show="alert.show"
-          :variant="alert.type"
-          fade
-          @click.native="triggerAlert(null)"
-          >{{ alert.msg }}</b-alert
-        >
         <div class="tx-contents">
           <div class="content-container mobile-hide">
             <interface-address
               :address="address"
-              :trigger-alert="triggerAlert"
               :print="print"
               :switch-addr="switchAddress"
               :qrcode="openAddressQrcode"
@@ -90,7 +81,6 @@
               :get-token-balance="getTokenBalance"
               :tokens="tokens"
               :received-tokens="receivedTokens"
-              :trigger-alert="triggerAlert"
             />
           </div>
         </div>
@@ -116,7 +106,7 @@ import InterfaceTokens from './components/InterfaceTokens';
 import MobileInterfaceAddress from './components/MobileInterfaceAddress';
 import PrintModal from './components/PrintModal';
 import { Web3Wallet } from '@/wallets/software';
-import { ErrorHandler } from '@/helpers';
+import { Toast } from '@/helpers';
 import { toChecksumAddress } from '@/helpers/addressUtils';
 import * as networkTypes from '@/networks/types';
 import { BigNumber } from 'bignumber.js';
@@ -124,6 +114,7 @@ import store from 'store';
 import TokenBalance from '@myetherwallet/eth-token-balance';
 import sortByBalance from '@/helpers/sortByBalance.js';
 import AddressQrcodeModal from '@/components/AddressQrcodeModal';
+import web3Utils from 'web3-utils';
 import {
   LedgerWallet,
   TrezorWallet,
@@ -195,12 +186,12 @@ export default {
     },
     ...mapGetters({
       network: 'network',
-      wallet: 'wallet',
       account: 'account',
       online: 'online',
       web3: 'web3',
       Networks: 'Networks',
-      sidemenuOpen: 'sidemenuOpen'
+      sidemenuOpen: 'sidemenuOpen',
+      wallet: 'wallet'
     })
   },
   watch: {
@@ -241,14 +232,18 @@ export default {
     switchAddress() {
       switch (this.account.identifier) {
         case LEDGER_TYPE:
-          LedgerWallet().then(_newWallet => {
-            this.toggleNetworkAddrModal(_newWallet);
-          });
+          LedgerWallet()
+            .then(_newWallet => {
+              this.toggleNetworkAddrModal(_newWallet);
+            })
+            .catch(LedgerWallet.errorHandler);
           break;
         case TREZOR_TYPE:
-          TrezorWallet().then(_newWallet => {
-            this.toggleNetworkAddrModal(_newWallet);
-          });
+          TrezorWallet()
+            .then(_newWallet => {
+              this.toggleNetworkAddrModal(_newWallet);
+            })
+            .catch(TrezorWallet.errorHandler);
           break;
         case BITBOX_TYPE:
           this.togglePasswordModal(BitBoxWallet, 'DigitalBitbox');
@@ -260,12 +255,14 @@ export default {
           this.$refs.mnemonicPhraseModal.$refs.mnemonicPhrase.show();
           break;
         case KEEPKEY_TYPE:
-          KeepkeyWallet(false, this.$eventHub).then(_newWallet => {
-            this.toggleNetworkAddrModal(_newWallet);
-          });
+          KeepkeyWallet(false, this.$eventHub)
+            .then(_newWallet => {
+              this.toggleNetworkAddrModal(_newWallet);
+            })
+            .catch(KeepkeyWallet.errorHandler);
           break;
         default:
-          ErrorHandler(
+          Toast.responseHandler(
             new Error(
               `Wallet type ${this.account.identifier} can't switch addresses`
             ),
@@ -276,34 +273,11 @@ export default {
     print() {
       this.$refs.printModal.$refs.print.show();
     },
-    triggerAlert(msg, type) {
-      let timeout;
-      if (msg !== null) {
-        this.alert = {
-          show: true,
-          msg: msg,
-          type: type ? type : 'info'
-        };
-        timeout = setTimeout(() => {
-          this.alert = {
-            show: false,
-            msg: '',
-            type: type ? type : 'info'
-          };
-        }, 3000);
-      } else {
-        clearTimeout(timeout);
-        this.alert = {
-          show: false,
-          msg: ''
-        };
-      }
-    },
     toggleSideMenu() {
       this.$store.commit('TOGGLE_SIDEMENU');
     },
     async fetchTokens() {
-      this.receivedTokens = true;
+      this.receivedTokens = false;
       let tokens = [];
       if (this.network.type.chainID === 1 || this.network.type.chainID === 3) {
         const tb = new TokenBalance(this.web3.currentProvider);
@@ -369,7 +343,7 @@ export default {
           return tokenBalance;
         })
         .catch(e => {
-          ErrorHandler(e, false);
+          Toast.responseHandler(e, false);
         });
       return balance;
     },
@@ -383,7 +357,7 @@ export default {
       store.set('customTokens', customTokenStore);
     },
     async setTokens() {
-      this.receivedTokens = false;
+      const customStore = store.get('customTokens');
       this.tokens = [];
       let tokens = await this.fetchTokens();
       tokens = tokens
@@ -413,21 +387,35 @@ export default {
         });
 
       this.tokens = tokens.sort(sortByBalance);
-      let customTokens = [];
+
       if (
-        store.get('customTokens') !== undefined &&
-        store.get('customTokens')[this.network.type.name] !== undefined &&
-        store.get('customTokens')[this.network.type.name].length > 0
+        customStore !== undefined &&
+        customStore[this.network.type.name] !== undefined &&
+        customStore[this.network.type.name].length > 0
       ) {
-        customTokens = store.get('customTokens')[
-          // eslint-disable-next-line
-          this.network.type.name
-        ].filter(token => token.balance > 0);
+        new Promise(resolve => {
+          const newArr = customStore[this.network.type.name].map(
+            async token => {
+              token.balance = await this.getTokenBalance(token);
+              return token;
+            }
+          );
+          Promise.all(newArr).then(res => {
+            customStore[this.network.type.name] = res;
+            store.set('customTokens', customStore);
+            resolve(res);
+          });
+        }).then(res => {
+          const allTokens = this.tokens
+            .filter(token => token.balance > 0)
+            .concat(res.filter(token => token.balance > 0));
+          this.tokensWithBalance = allTokens;
+          this.receivedTokens = true;
+        });
+      } else {
+        this.receivedTokens = true;
+        this.tokensWithBalance = this.tokens.filter(token => token.balance > 0);
       }
-      const allTokens = this.tokens
-        .filter(token => token.balance > 0)
-        .concat(customTokens);
-      this.tokensWithBalance = allTokens;
     },
     getBlock() {
       this.web3.eth
@@ -436,7 +424,7 @@ export default {
           this.blockNumber = res;
         })
         .catch(e => {
-          ErrorHandler(e, false);
+          Toast.responseHandler(e, Toast.ERROR);
         });
     },
     getBalance() {
@@ -448,21 +436,24 @@ export default {
           this.$store.dispatch('setAccountBalance', res);
         })
         .catch(e => {
-          ErrorHandler(e, false);
+          Toast.responseHandler(e, false);
         });
     },
     checkWeb3WalletAddrChange() {
       this.pollAddress = setInterval(() => {
         window.web3.eth.getAccounts((err, accounts) => {
           if (err) {
-            return ErrorHandler(err, false);
+            return Toast.responseHandler(err, false);
           }
           if (!accounts.length) {
-            return ErrorHandler(new Error('Please unlock metamask'), false);
+            return Toast.responseHandler(
+              new Error('Please unlock metamask'),
+              Toast.ERROR
+            );
           }
           const address = accounts[0];
           if (
-            this.wallet !== null &&
+            this.account.address !== null &&
             address.toLowerCase() !== this.account.address.toLowerCase()
           ) {
             const wallet = new Web3Wallet(address);
@@ -477,6 +468,7 @@ export default {
     },
     matchWeb3WalletNetwork() {
       this.pollNetwork = setInterval(() => {
+        if (!window.web3.eth.net) return;
         window.web3.eth.net
           .getId()
           .then(netId => {
@@ -491,7 +483,7 @@ export default {
             }
           })
           .catch(e => {
-            ErrorHandler(e, false);
+            Toast.responseHandler(e, false);
           });
       }, 500);
     },
@@ -500,7 +492,7 @@ export default {
       clearInterval(this.pollBlock);
       clearInterval(this.pollAddress);
     },
-    setupOnlineEnvironment() {
+    setupOnlineEnvironment: web3Utils._.debounce(function() {
       this.clearIntervals();
       if (store.get('customTokens') === undefined) {
         store.set('customTokens', {});
@@ -509,7 +501,7 @@ export default {
         this.setCustomTokenStore();
       }
       if (this.online === true) {
-        if (this.wallet !== null) {
+        if (this.account.address !== null) {
           if (this.account.identifier === WEB3_TYPE) {
             this.checkWeb3WalletAddrChange();
             this.matchWeb3WalletNetwork();
@@ -523,7 +515,7 @@ export default {
           this.getHighestGas();
         }
       }
-    },
+    }),
     getHighestGas() {
       this.web3.eth
         .getGasPrice()
@@ -533,7 +525,7 @@ export default {
           ).toNumber();
         })
         .catch(e => {
-          ErrorHandler(e, false);
+          Toast.responseHandler(e, true);
         });
     },
     setENS() {