+++ src/layouts/InterfaceLayout/components/InterfaceTokens/InterfaceTokens.vue
@@ -20,7 +20,7 @@
           </div>
         </div>
         <div ref="tokenTableContainer" class="token-table-container">
-          <table v-show="customTokens.length > 0">
+          <table v-show="customTokens.length > 0 && receivedTokens">
             <tr
               v-for="(token, index) in customTokens"
               :key="token.name + index"
@@ -36,7 +36,7 @@
             </tr>
           </table>
 
-          <table v-show="localTokens.length > 0">
+          <table v-show="localTokens.length > 0 && receivedTokens">
             <tr v-for="(token, index) in localTokens" :key="token.name + index">
               <td>{{ token.name }}</td>
               <td
@@ -51,7 +51,9 @@
           </table>
 
           <div
-            v-show="search === '' && localTokens.length === 0 && receivedTokens"
+            v-show="
+              search === '' && localTokens.length === 0 && !receivedTokens
+            "
             class="spinner-container"
           >
             <i class="fa fa-spinner fa-spin" />
@@ -92,7 +94,7 @@
 <script>
 import store from 'store';
 import { mapGetters } from 'vuex';
-import { ErrorHandler } from '@/helpers';
+import { Toast } from '@/helpers';
 import InterfaceTokensModal from '../InterfaceTokensModal';
 import sortByBalance from '@/helpers/sortByBalance.js';
 import utils from 'web3-utils';
@@ -117,10 +119,6 @@ export default {
       type: Function,
       default: function() {}
     },
-    triggerAlert: {
-      type: Function,
-      default: function() {}
-    },
     fetchTokens: {
       type: Function,
       default: function() {}
@@ -142,28 +140,17 @@ export default {
     })
   },
   watch: {
+    receivedTokens() {
+      this.getCustomTokens();
+    },
     tokens(newVal) {
       this.assignTokens(newVal, this.search);
+      this.getCustomTokens();
     },
     search(newVal) {
       this.assignTokens(this.tokens, newVal);
-    },
-    network(newVal) {
-      if (
-        store.get('customTokens') !== undefined &&
-        store.get('customTokens')[newVal.type.name] !== undefined
-      ) {
-        this.customTokens = store.get('customTokens')[newVal.type.name];
-      } else {
-        this.customTokens = [];
-      }
-    }
-  },
-  mounted() {
-    if (store.get('customTokens')) {
       this.getCustomTokens();
     }
-    this.assignTokens(this.tokens, this.search);
   },
   methods: {
     getV3Tokens() {
@@ -215,6 +202,7 @@ export default {
       this.customTokens.splice(idx, 1);
       storedTokens[this.network.type.name] = this.customTokens;
       store.set('customTokens', storedTokens);
+      this.fetchTokens();
     },
     searchBySymbol(symbol) {
       const searchNetwork = this.localTokens.find(item => {
@@ -255,21 +243,20 @@ export default {
       const findTokenByAddr = this.searchByAddr(address);
       if (!findTokenByAddr && addType !== '') {
         this.$refs.tokenModal.$refs.token.hide();
-        this.triggerAlert(
-          'A default token with this contract address already exists!',
-          'danger'
+        Toast.responseHandler(
+          'A default or custom token with this contract address already exists!',
+          Toast.ERROR
         );
         return false;
       } else if (!findTokenBySymbol && addType !== '') {
         this.$refs.tokenModal.$refs.token.hide();
-        this.triggerAlert(
-          "A default token with this symbol already exists! The token in our list may have the same symbol but a different contract address, try adding it again with a '2' after the symbol!",
-          'danger'
+        Toast.responseHandler(
+          "A default or custom token with this symbol already exists! The token in our list may have the same symbol but a different contract address, try adding it again with a '2' after the symbol!",
+          Toast.ERROR
         );
         return false;
       }
-
-      return !findTokenByAddr || !findTokenBySymbol;
+      return findTokenByAddr || findTokenBySymbol;
     },
     async addToken(address, symbol, decimal) {
       if (this.tokenError(address, symbol, 'manual')) {
@@ -287,13 +274,17 @@ export default {
           this.customTokens.length > 0 ? this.customTokens : [];
         token['balance'] = await this.getTokenBalance(token);
         if (token['balance'] === undefined) {
-          ErrorHandler(new Error('Token Balance Returned Undefined'), false);
+          Toast.responseHandler(
+            new Error('Token Balance Returned Undefined'),
+            Toast.ERROR
+          );
         }
         this.customTokens.push(token);
         currentCustomToken[this.network.type.name] = this.customTokens;
         store.set('customTokens', currentCustomToken);
         this.$refs.tokenModal.$refs.token.hide();
-        this.triggerAlert('Successfully added token!');
+        Toast.responseHandler('Successfully added token!', Toast.SUCCESS);
+        this.fetchTokens();
       }
     },
     tokenListExpend() {
@@ -302,7 +293,7 @@ export default {
       this.$refs.expendUp.classList.toggle('hidden');
     },
     async assignTokens(arr, query) {
-      const oldArray = this.customTokens.slice();
+      const oldArray = this.customTokens ? this.customTokens.slice() : [];
       if (query !== '') {
         this.customTokens = oldArray
           .filter(token => {
@@ -320,9 +311,6 @@ export default {
           .sort(sortByBalance);
       } else {
         this.localTokens = arr;
-        if (store.get('customTokens') !== undefined) {
-          this.customTokens = store.get('customTokens')[this.network.type.name];
-        }
       }
     }
   }