+++ src/layouts/InterfaceLayout/containers/SendCurrencyContainer/SendCurrencyContainer.vue
@@ -47,7 +47,7 @@
               v-if="!isValidAmount || errors.has('value')"
               class="error-message-container"
             >
-              <p>{{ $t('common.dontHaveEnough') }}</p>
+              <p>{{ $t('common.notAValidAmount') }}</p>
             </div>
           </div>
         </div>
@@ -181,7 +181,7 @@ import CurrencyPicker from '../../components/CurrencyPicker';
 import InterfaceBottomText from '@/components/InterfaceBottomText';
 import Blockie from '@/components/Blockie';
 import EthTx from 'ethereumjs-tx';
-import { Misc, ErrorHandler } from '@/helpers';
+import { Misc, Toast } from '@/helpers';
 import BigNumber from 'bignumber.js';
 import ethUnit from 'ethjs-unit';
 import utils from 'web3-utils';
@@ -237,10 +237,22 @@ export default {
       if (this.isToken) {
         return (
           new BigNumber(this.value).lte(this.selectedCurrency.balance) &&
-          new BigNumber(txFeeEth).lte(this.balanceDefault)
+          new BigNumber(txFeeEth).lte(this.balanceDefault) &&
+          this.isValidDecimals
         );
       }
-      return new BigNumber(this.value).plus(txFeeEth).lte(this.balanceDefault);
+      return (
+        new BigNumber(this.value).plus(txFeeEth).lte(this.balanceDefault) &&
+        this.isValidDecimals
+      );
+    },
+    isValidDecimals() {
+      const decimals = (this.value + '').split('.')[1];
+      if (!decimals) return true;
+      if (this.isToken) {
+        return decimals.length <= this.selectedCurrency.decimals;
+      }
+      return decimals.length <= 18;
     },
     isValidData() {
       return Misc.validateHexString(this.data);
@@ -352,7 +364,7 @@ export default {
         })
         .catch(err => {
           this.gasLimit = -1;
-          ErrorHandler(err, true);
+          Toast.responseHandler(err, Toast.ERROR);
         });
     },
     async submitTransaction() {
@@ -374,10 +386,10 @@ export default {
         const json = _tx.toJSON(true);
         json.from = coinbase;
         this.web3.eth.sendTransaction(json).catch(err => {
-          ErrorHandler(err, true);
+          Toast.responseHandler(err, Toast.ERROR);
         });
       } catch (e) {
-        throw e;
+        Toast.responseHandler(e, Toast.ERROR);
       }
     },
     copyToClipboard(ref) {