+++ src/partners/bity/bity.js
@@ -7,7 +7,7 @@ import {
   swapNotificationStatuses
 } from '../partnersConfig';
 import { utils } from '../helpers';
-import { ErrorHandler } from '@/helpers';
+import { Toast } from '@/helpers';
 import {
   getRates,
   openOrder,
@@ -330,11 +330,13 @@ export default class BitySwap {
           swapDetails.providerAddress =
             swapDetails.dataForInitialization.payment_address;
           swapDetails.isDex = BitySwap.isDex();
+        } else {
+          throw Error('abort');
         }
       }
     } else if (!this.checkIfExit(swapDetails)) {
       swapDetails.dataForInitialization = await this.buildOrder(swapDetails);
-      if (!swapDetails.dataForInitialization) throw Error('invalid');
+      if (!swapDetails.dataForInitialization) throw Error('abort');
       swapDetails.providerReceives =
         swapDetails.dataForInitialization.input.amount;
       swapDetails.providerSends =
@@ -436,7 +438,7 @@ export default class BitySwap {
       };
       return buildCyptoToFiatOrderData(orderData);
     } catch (e) {
-      ErrorHandler(e, false);
+      Toast.responseHandler(e, false);
     }
   }
 
@@ -462,7 +464,7 @@ export default class BitySwap {
   static parseExitOrder(order) {
     return {
       orderId: order.id,
-      statusId: order.id,
+      statusId: order.reference,
       sendToAddress: order.payment_address,
       recValue: order.amount,
       sendValue: order.payment_amount,
@@ -481,7 +483,7 @@ export default class BitySwap {
 
   static async getOrderStatusCrypto(noticeDetails) {
     try {
-      const data = await getStatus(noticeDetails.statusId);
+      const data = await getStatus(noticeDetails.orderId);
       if (data.status === bityStatuses.EXEC) {
         return swapNotificationStatuses.COMPLETE;
       }
@@ -506,7 +508,7 @@ export default class BitySwap {
         }
       }
     } catch (e) {
-      ErrorHandler(e, false);
+      Toast.responseHandler(e, false);
     }
     return swapNotificationStatuses.PENDING;
   }
@@ -514,7 +516,7 @@ export default class BitySwap {
   static async getOrderStatusFiat(noticeDetails) {
     try {
       const data = await getStatusFiat(
-        noticeDetails.statusId,
+        noticeDetails.orderId,
         noticeDetails.special
       );
       if (!utils.isJson(data)) return swapNotificationStatuses.PENDING;
@@ -545,7 +547,7 @@ export default class BitySwap {
           return swapNotificationStatuses.CANCELLED;
       }
     } catch (e) {
-      ErrorHandler(e, false);
+      Toast.responseHandler(e, false);
     }
     return swapNotificationStatuses.PENDING;
   }