+++ src/router/index.js
@@ -1,10 +1,6 @@
-import Vue from 'vue';
 import Router from 'vue-router';
 import store from '@/store';
-
 import { getMode, getRoutes } from '@/builds/configs';
-
-Vue.use(Router);
 
 const router = new Router({
   mode: getMode(),